Fix useless-cast warnings

+++ include/cereal/types/string.hpp
@@ -52,8 +52,8 @@ namespace cereal
   {
     size_type size;
     ar( make_size_tag( size ) );
-    str.resize(static_cast<std::size_t>(size));
-    ar( binary_data( const_cast<CharT *>( str.data() ), static_cast<std::size_t>(size) * sizeof(CharT) ) );
+    str.resize(size);
+    ar( binary_data( const_cast<CharT *>( str.data() ), size * sizeof(CharT) ) );
   }
 } // namespace cereal
